* new:[API] lav< { int8_t, int16_t, int32_t } > With native bounds * fix:[API] lav< { uint8_t, uint16_t, uint32_t } > Use native scale i.e scale compare to native is 1 * enhance:[internal] Small internal clean of on unnecessary things * fix:[internal] make::gray use proper type conversion * fix:[internal] get:chroma( lab )

+++ src/color/generic/trait/scalar.hpp
@@ -13,10 +13,10 @@ namespace color
 
     template< typename category_name >
      struct scalar
-      : public ::color::_internal::utility::type::traitC< double >
+      : public ::color::_internal::utility::type::traitC< long double >
       {
        public:
-        typedef double instance_type;
+        typedef long double instance_type;
 
         static bool is_small( instance_type const& value )
          {
